+++ transfer_sync_service/lambda/get_list_status/get_list_status.py
@@ -1,5 +1,4 @@
 import boto3
-import botocore
 import json
 import copy
 from aws_lambda_powertools import Logger